Move AnyMap from EnvironmentManager into separate file
