+++ controller/Api/Api.csproj
@@ -79,6 +79,7 @@
     <Compile Include="Runtime\ApiProvider.cs" />
     <Compile Include="Runtime\IApiProvider.cs" />
     <Compile Include="UIManager\UIManager.cs" />
+    <Compile Include="Utilities\AnyMap.cs" />
     <Compile Include="VirtualFileSystem\DirectroyExpection.cs" />
     <Compile Include="VirtualFileSystem\ExposedApiAttribute.cs" />
     <Compile Include="Utilities\ParamsHelper.cs" />
